changed to async await, granulation and format
