+++ src/App.jsx
@@ -1,9 +1,6 @@
 import { useEffect, useState } from 'react'
-import reactLogo from './assets/react.svg'
-import viteLogo from '/vite.svg'
 import './App.css'
-import { useArtworks } from './hooks/useArtworks'
-import { Artworks } from './components/Artworks'
+import Artworks from './components/Artworks'
 import { getArtworks } from './services/artworks.js'
 
 function App() {
@@ -13,11 +10,8 @@ function App() {
     getArtworks().then(res => setArtworks(res)).catch(error => console.error("Error ", error))
   }, [])
   
-  //const artworks = useArtworks()
 
-  //const aux = getArtworks().then(res => console.log(res))
 
-  console.log('artworks: ', artworks)
   return (
     <div className='page'>
       <main>
